refactor: split phase 1 oversized files
