--- internal/core/model/model.go
@@ -1,369 +1,0 @@
-package model
-
-import (
-	"context"
-	"path/filepath"
-	"strings"
-	"time"
-
-	"github.com/compozy/compozy/internal/core/run/journal"
-)
-
-const (
-	UnknownFileName          = "unknown"
-	IDECodex                 = "codex"
-	IDEClaude                = "claude"
-	IDEDroid                 = "droid"
-	IDECursor                = "cursor-agent"
-	IDEOpenCode              = "opencode"
-	IDEPi                    = "pi"
-	IDEGemini                = "gemini"
-	IDECopilot               = "copilot"
-	DefaultCodexModel        = "gpt-5.4"
-	DefaultClaudeModel       = "opus"
-	DefaultCursorModel       = "composer-1"
-	DefaultOpenCodeModel     = "anthropic/claude-opus-4-6"
-	DefaultPiModel           = "anthropic/claude-opus-4-6"
-	DefaultGeminiModel       = "gemini-2.5-pro"
-	DefaultCopilotModel      = "claude-sonnet-4.6"
-	DefaultActivityTimeout   = 10 * time.Minute
-	WorkflowRootDirName      = ".compozy"
-	WorkflowConfigFileName   = "config.toml"
-	WorkflowTasksDirName     = "tasks"
-	WorkflowRunsDirName      = "runs"
-	ArchivedWorkflowDirName  = "_archived"
-	ModeCodeReview           = "pr-review"
-	ModePRDTasks             = "prd-tasks"
-	ModeExec                 = "exec"
-	AccessModeDefault        = "default"
-	AccessModeFull           = "full"
-	OutputFormatTextValue    = "text"
-	OutputFormatJSONValue    = "json"
-	OutputFormatRawJSONValue = "raw-json"
-)
-
-type ExecutionMode string
-
-const (
-	ExecutionModePRReview ExecutionMode = ModeCodeReview
-	ExecutionModePRDTasks ExecutionMode = ModePRDTasks
-	ExecutionModeExec     ExecutionMode = ModeExec
-)
-
-type OutputFormat string
-
-const (
-	OutputFormatText    OutputFormat = OutputFormatTextValue
-	OutputFormatJSON    OutputFormat = OutputFormatJSONValue
-	OutputFormatRawJSON OutputFormat = OutputFormatRawJSONValue
-)
-
-type RuntimeConfig struct {
-	WorkspaceRoot          string
-	Name                   string
-	Round                  int
-	Provider               string
-	PR                     string
-	ReviewsDir             string
-	TasksDir               string
-	DryRun                 bool
-	AutoCommit             bool
-	Concurrent             int
-	BatchSize              int
-	IDE                    string
-	Model                  string
-	AddDirs                []string
-	TailLines              int
-	ReasoningEffort        string
-	AccessMode             string
-	Mode                   ExecutionMode
-	OutputFormat           OutputFormat
-	Verbose                bool
-	TUI                    bool
-	Persist                bool
-	RunID                  string
-	PromptText             string
-	PromptFile             string
-	ReadPromptStdin        bool
-	ResolvedPromptText     string
-	IncludeCompleted       bool
-	IncludeResolved        bool
-	Timeout                time.Duration
-	MaxRetries             int
-	RetryBackoffMultiplier float64
-}
-
-func (cfg *RuntimeConfig) ApplyDefaults() {
-	if cfg.Concurrent <= 0 {
-		cfg.Concurrent = 1
-	}
-	if cfg.BatchSize <= 0 {
-		cfg.BatchSize = 1
-	}
-	if cfg.IDE == "" {
-		cfg.IDE = IDECodex
-	}
-	if cfg.TailLines < 0 {
-		cfg.TailLines = 0
-	}
-	if cfg.ReasoningEffort == "" {
-		cfg.ReasoningEffort = "medium"
-	}
-	if cfg.AccessMode == "" {
-		cfg.AccessMode = AccessModeFull
-	}
-	if cfg.Mode == "" {
-		cfg.Mode = ExecutionModePRReview
-	}
-	if cfg.OutputFormat == "" {
-		cfg.OutputFormat = OutputFormatText
-	}
-	if cfg.Timeout <= 0 {
-		cfg.Timeout = DefaultActivityTimeout
-	}
-	if cfg.RetryBackoffMultiplier <= 0 {
-		cfg.RetryBackoffMultiplier = 1.5
-	}
-}
-
-func TasksBaseDir() string {
-	return TasksBaseDirForWorkspace("")
-}
-
-func TaskDirectory(name string) string {
-	return TaskDirectoryForWorkspace("", name)
-}
-
-func CompozyDir(workspaceRoot string) string {
-	trimmed := strings.TrimSpace(workspaceRoot)
-	if trimmed == "" {
-		return WorkflowRootDirName
-	}
-	return filepath.Join(filepath.Clean(trimmed), WorkflowRootDirName)
-}
-
-func ConfigPathForWorkspace(workspaceRoot string) string {
-	return filepath.Join(CompozyDir(workspaceRoot), WorkflowConfigFileName)
-}
-
-func TasksBaseDirForWorkspace(workspaceRoot string) string {
-	return filepath.Join(CompozyDir(workspaceRoot), WorkflowTasksDirName)
-}
-
-func RunsBaseDirForWorkspace(workspaceRoot string) string {
-	return filepath.Join(CompozyDir(workspaceRoot), WorkflowRunsDirName)
-}
-
-func TaskDirectoryForWorkspace(workspaceRoot, name string) string {
-	return filepath.Join(TasksBaseDirForWorkspace(workspaceRoot), name)
-}
-
-type RunArtifacts struct {
-	RunID       string
-	RunDir      string
-	RunMetaPath string
-	EventsPath  string
-	TurnsDir    string
-	JobsDir     string
-	ResultPath  string
-}
-
-type JobArtifacts struct {
-	PromptPath string
-	OutLogPath string
-	ErrLogPath string
-}
-
-func NewRunArtifacts(workspaceRoot, runID string) RunArtifacts {
-	safeRunID := sanitizeRunID(runID)
-	runDir := filepath.Join(RunsBaseDirForWorkspace(workspaceRoot), safeRunID)
-	return RunArtifacts{
-		RunID:       safeRunID,
-		RunDir:      runDir,
-		RunMetaPath: filepath.Join(runDir, "run.json"),
-		EventsPath:  filepath.Join(runDir, "events.jsonl"),
-		TurnsDir:    filepath.Join(runDir, "turns"),
-		JobsDir:     filepath.Join(runDir, "jobs"),
-		ResultPath:  filepath.Join(runDir, "result.json"),
-	}
-}
-
-func sanitizeRunID(runID string) string {
-	trimmed := strings.TrimSpace(runID)
-	if trimmed == "" {
-		return "run"
-	}
-	normalized := strings.NewReplacer("/", "-", "\\", "-").Replace(trimmed)
-
-	var builder strings.Builder
-	builder.Grow(len(normalized))
-	lastDash := false
-	for _, r := range normalized {
-		switch {
-		case r >= 'a' && r <= 'z', r >= 'A' && r <= 'Z', r >= '0' && r <= '9', r == '.', r == '_':
-			builder.WriteRune(r)
-			lastDash = false
-		case r == '-':
-			builder.WriteRune(r)
-			lastDash = true
-		default:
-			if lastDash {
-				continue
-			}
-			builder.WriteByte('-')
-			lastDash = true
-		}
-	}
-
-	safe := strings.Trim(builder.String(), "-")
-	switch safe {
-	case "", ".", "..":
-		return "run"
-	default:
-		return safe
-	}
-}
-
-func (artifacts RunArtifacts) JobArtifacts(safeName string) JobArtifacts {
-	return JobArtifacts{
-		PromptPath: filepath.Join(artifacts.JobsDir, safeName+".prompt.md"),
-		OutLogPath: filepath.Join(artifacts.JobsDir, safeName+".out.log"),
-		ErrLogPath: filepath.Join(artifacts.JobsDir, safeName+".err.log"),
-	}
-}
-
-func ArchivedTasksDir(baseDir string) string {
-	return filepath.Join(baseDir, ArchivedWorkflowDirName)
-}
-
-func IsActiveWorkflowDirName(name string) bool {
-	trimmed := strings.TrimSpace(name)
-	return trimmed != "" && !strings.HasPrefix(trimmed, ".") && trimmed != ArchivedWorkflowDirName
-}
-
-type IssueEntry struct {
-	Name     string
-	AbsPath  string
-	Content  string
-	CodeFile string
-}
-
-type ReviewContext struct {
-	Status      string
-	File        string
-	Line        int
-	Severity    string
-	Author      string
-	ProviderRef string
-}
-
-type RoundMeta struct {
-	Provider   string
-	PR         string
-	Round      int
-	CreatedAt  time.Time
-	Total      int
-	Resolved   int
-	Unresolved int
-}
-
-type TaskMeta struct {
-	CreatedAt time.Time
-	UpdatedAt time.Time
-	Total     int
-	Completed int
-	Pending   int
-}
-
-type TaskEntry struct {
-	Content      string
-	Status       string
-	Title        string
-	TaskType     string
-	Complexity   string
-	Dependencies []string
-}
-
-type TaskFileMeta struct {
-	Status       string   `yaml:"status"`
-	Title        string   `yaml:"title"`
-	TaskType     string   `yaml:"type"`
-	Complexity   string   `yaml:"complexity,omitempty"`
-	Dependencies []string `yaml:"dependencies"`
-}
-
-type ReviewFileMeta struct {
-	Status      string `yaml:"status"`
-	File        string `yaml:"file,omitempty"`
-	Line        int    `yaml:"line,omitempty"`
-	Severity    string `yaml:"severity,omitempty"`
-	Author      string `yaml:"author,omitempty"`
-	ProviderRef string `yaml:"provider_ref,omitempty"`
-}
-
-type JournalHandle interface {
-	Journal() *journal.Journal
-	Close(context.Context) error
-}
-
-type SolvePreparation struct {
-	Jobs         []Job
-	RunArtifacts RunArtifacts
-	// JournalHandle carries the run journal while keeping cleanup ownership
-	// explicit. Kernel/runtime flows retain responsibility for closing it.
-	JournalHandle    JournalHandle
-	InputDir         string
-	InputDirPath     string
-	ResolvedName     string
-	ResolvedPR       string
-	ResolvedProvider string
-	ResolvedRound    int
-}
-
-func (p *SolvePreparation) Journal() *journal.Journal {
-	if p == nil || p.JournalHandle == nil {
-		return nil
-	}
-	return p.JournalHandle.Journal()
-}
-
-func (p *SolvePreparation) SetJournal(j *journal.Journal) {
-	if p == nil {
-		return
-	}
-	if j == nil {
-		p.JournalHandle = nil
-		return
-	}
-	p.JournalHandle = journal.NewOwner(j)
-}
-
-func (p *SolvePreparation) CloseJournal(ctx context.Context) error {
-	if p == nil || p.JournalHandle == nil {
-		return nil
-	}
-	handle := p.JournalHandle
-	p.JournalHandle = nil
-	return handle.Close(ctx)
-}
-
-type Job struct {
-	CodeFiles     []string
-	Groups        map[string][]IssueEntry
-	TaskTitle     string
-	TaskType      string
-	SafeName      string
-	Prompt        []byte
-	SystemPrompt  string
-	OutPromptPath string
-	OutLog        string
-	ErrLog        string
-}
-
-func (j Job) IssueCount() int {
-	total := 0
-	for _, items := range j.Groups {
-		total += len(items)
-	}
-	return total
-}
