+++ internal/core/run/lifecycle.go
@@ -1,0 +1,414 @@
+package run
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"strings"
+	"sync/atomic"
+	"time"
+
+	"github.com/compozy/compozy/internal/core/model"
+	"github.com/compozy/compozy/internal/core/run/journal"
+	"github.com/compozy/compozy/internal/core/tasks"
+	"github.com/compozy/compozy/pkg/compozy/events"
+	"github.com/compozy/compozy/pkg/compozy/events/kinds"
+)
+
+type jobLifecycle struct {
+	index          int
+	job            *job
+	execCtx        *jobExecutionContext
+	state          jobPhase
+	attempt        int
+	startedAt      time.Time
+	currentTimeout time.Duration
+	lastExitCode   int
+	lastFailure    *failInfo
+}
+
+func newJobLifecycle(index int, jb *job, execCtx *jobExecutionContext) *jobLifecycle {
+	return &jobLifecycle{
+		index:   index,
+		job:     jb,
+		execCtx: execCtx,
+		state:   jobPhaseQueued,
+	}
+}
+
+func (l *jobLifecycle) schedule() {
+	l.state = jobPhaseScheduled
+}
+
+func (l *jobLifecycle) startAttempt(attempt int, maxAttempts int, timeout time.Duration) {
+	l.attempt = attempt
+	l.currentTimeout = timeout
+	l.state = jobPhaseRunning
+	if l.startedAt.IsZero() {
+		l.startedAt = time.Now().UTC()
+	}
+	cfg := l.execConfig()
+	if l.attempt == 1 {
+		l.execCtx.submitEventOrWarn(
+			events.EventKindJobStarted,
+			kinds.JobStartedPayload{
+				JobAttemptInfo: kinds.JobAttemptInfo{
+					Index:       l.index,
+					Attempt:     attempt,
+					MaxAttempts: maxAttempts,
+				},
+			},
+		)
+		notifyJobStart(
+			cfg != nil && cfg.humanOutputEnabled() && l.execCtx.ui == nil,
+			l.job,
+			l.configIDE(),
+			l.configModel(),
+			l.configAddDirs(),
+			l.configReasoningEffort(),
+			l.configAccessMode(),
+		)
+	}
+}
+
+func (l *jobLifecycle) markRetry(failure failInfo, nextAttempt int, maxAttempts int) {
+	l.lastFailure = &failure
+	l.lastExitCode = failure.exitCode
+	l.state = jobPhaseRetrying
+	l.attempt = nextAttempt
+	l.job.exitCode = failure.exitCode
+	l.job.failure = failure.err.Error()
+	l.execCtx.submitEventOrWarn(
+		events.EventKindJobRetryScheduled,
+		kinds.JobRetryScheduledPayload{
+			JobAttemptInfo: kinds.JobAttemptInfo{
+				Index:       l.index,
+				Attempt:     nextAttempt,
+				MaxAttempts: maxAttempts,
+			},
+			Reason: failure.err.Error(),
+		},
+	)
+}
+
+func (l *jobLifecycle) markGiveUp(failure failInfo) {
+	l.lastFailure = &failure
+	l.lastExitCode = failure.exitCode
+	l.state = jobPhaseFailed
+	l.job.status = runStatusFailed
+	l.job.exitCode = failure.exitCode
+	l.job.failure = failure.err.Error()
+	if l.lastFailure != nil {
+		recordFailure(&l.execCtx.failuresMu, &l.execCtx.failures, *l.lastFailure)
+	}
+	atomic.AddInt32(&l.execCtx.failed, 1)
+	l.execCtx.submitEventOrWarn(
+		events.EventKindJobFailed,
+		kinds.JobFailedPayload{
+			JobAttemptInfo: kinds.JobAttemptInfo{
+				Index:       l.index,
+				Attempt:     l.attempt,
+				MaxAttempts: l.maxAttempts(),
+			},
+			CodeFile: l.job.codeFileLabel(),
+			ExitCode: l.lastExitCode,
+			OutLog:   l.job.outLog,
+			ErrLog:   l.job.errLog,
+			Error:    l.job.failure,
+		},
+	)
+	if l.lastFailure != nil && l.humanOutputEnabled() {
+		fmt.Fprintf(
+			os.Stderr,
+			"\n❌ Job %d (%s) failed with exit code %d: %v\n",
+			l.index+1,
+			l.job.codeFileLabel(),
+			l.lastExitCode,
+			l.lastFailure.err,
+		)
+	}
+}
+
+func (l *jobLifecycle) markSuccess() {
+	if l.startedAt.IsZero() {
+		l.startedAt = time.Now().UTC()
+	}
+	l.lastFailure = nil
+	l.lastExitCode = 0
+	l.state = jobPhaseSucceeded
+	l.job.status = runStatusSucceeded
+	l.job.exitCode = 0
+	l.job.failure = ""
+	l.execCtx.submitEventOrWarn(
+		events.EventKindJobCompleted,
+		kinds.JobCompletedPayload{
+			JobAttemptInfo: kinds.JobAttemptInfo{
+				Index:       l.index,
+				Attempt:     l.attempt,
+				MaxAttempts: l.maxAttempts(),
+			},
+			ExitCode:   0,
+			DurationMs: time.Since(l.startedAt).Milliseconds(),
+		},
+	)
+}
+
+func (l *jobLifecycle) markCanceled(exitCode int) {
+	l.lastExitCode = exitCode
+	l.state = jobPhaseCanceled
+	l.job.status = runStatusCanceled
+	l.job.exitCode = exitCode
+	if exitCode == exitCodeCanceled {
+		l.lastFailure = &failInfo{
+			codeFile: l.job.codeFileLabel(),
+			exitCode: exitCodeCanceled,
+			outLog:   l.job.outLog,
+			errLog:   l.job.errLog,
+			err:      fmt.Errorf("job canceled by shutdown"),
+		}
+	} else {
+		l.lastFailure = nil
+	}
+	if l.lastFailure != nil {
+		l.job.failure = l.lastFailure.err.Error()
+	}
+
+	if l.lastFailure != nil {
+		recordFailure(&l.execCtx.failuresMu, &l.execCtx.failures, *l.lastFailure)
+	}
+	atomic.AddInt32(&l.execCtx.failed, 1)
+	reason := ""
+	if l.lastFailure != nil && l.lastFailure.err != nil {
+		reason = l.lastFailure.err.Error()
+	}
+	l.execCtx.submitEventOrWarn(
+		events.EventKindJobCancelled,
+		kinds.JobCancelledPayload{
+			JobAttemptInfo: kinds.JobAttemptInfo{
+				Index:       l.index,
+				Attempt:     l.attempt,
+				MaxAttempts: l.maxAttempts(),
+			},
+			Reason: reason,
+		},
+	)
+	if l.lastFailure != nil && l.humanOutputEnabled() {
+		fmt.Fprintf(
+			os.Stderr,
+			"\n⚠️ Job %d (%s) canceled: %v\n",
+			l.index+1,
+			l.job.codeFileLabel(),
+			l.lastFailure.err,
+		)
+	}
+}
+
+func (l *jobLifecycle) execConfig() *config {
+	if l == nil || l.execCtx == nil {
+		return nil
+	}
+	return l.execCtx.cfg
+}
+
+func (l *jobLifecycle) maxAttempts() int {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return 1
+	}
+	return atLeastOne(cfg.maxRetries + 1)
+}
+
+func (l *jobLifecycle) humanOutputEnabled() bool {
+	cfg := l.execConfig()
+	return cfg != nil && cfg.humanOutputEnabled()
+}
+
+func (l *jobLifecycle) configIDE() string {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return ""
+	}
+	return cfg.ide
+}
+
+func (l *jobLifecycle) configModel() string {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return ""
+	}
+	return cfg.model
+}
+
+func (l *jobLifecycle) configAddDirs() []string {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return nil
+	}
+	return cfg.addDirs
+}
+
+func (l *jobLifecycle) configReasoningEffort() string {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return ""
+	}
+	return cfg.reasoningEffort
+}
+
+func (l *jobLifecycle) configAccessMode() string {
+	cfg := l.execConfig()
+	if cfg == nil {
+		return ""
+	}
+	return cfg.accessMode
+}
+
+func refreshTaskMetaOnExit(cfg *config) {
+	if cfg == nil || cfg.mode != model.ExecutionModePRDTasks || strings.TrimSpace(cfg.tasksDir) == "" {
+		return
+	}
+
+	meta, err := tasks.RefreshTaskMeta(cfg.tasksDir)
+	if err != nil {
+		runtimeLoggerFor(cfg, cfg != nil && cfg.uiEnabled()).Warn(
+			"failed to refresh task workflow metadata at command exit",
+			"tasks_dir",
+			cfg.tasksDir,
+			"error",
+			err,
+		)
+		return
+	}
+
+	runtimeLoggerFor(cfg, cfg != nil && cfg.uiEnabled()).Info(
+		"refreshed task workflow metadata at command exit",
+		"tasks_dir",
+		cfg.tasksDir,
+		"completed",
+		meta.Completed,
+		"pending",
+		meta.Pending,
+		"total",
+		meta.Total,
+	)
+}
+
+func submitRunEvent(
+	ctx context.Context,
+	runJournal *journal.Journal,
+	runID string,
+	kind events.EventKind,
+	payload any,
+) error {
+	if runJournal == nil {
+		return nil
+	}
+	if ctx == nil {
+		ctx = context.Background()
+	}
+	event, err := newRuntimeEvent(runID, kind, payload)
+	if err != nil {
+		return err
+	}
+	return runJournal.Submit(ctx, event)
+}
+
+func emitRunTerminalEvent(
+	ctx context.Context,
+	runJournal *journal.Journal,
+	result executionResult,
+	jobs []job,
+	startedAt time.Time,
+) error {
+	if runJournal == nil {
+		return nil
+	}
+
+	var (
+		succeeded int
+		failed    int
+		canceled  int
+	)
+	for idx := range jobs {
+		switch jobs[idx].status {
+		case runStatusSucceeded:
+			succeeded++
+		case runStatusCanceled:
+			canceled++
+		case runStatusFailed:
+			failed++
+		}
+	}
+
+	durationMs := time.Since(startedAt).Milliseconds()
+	switch result.Status {
+	case runStatusSucceeded:
+		return submitRunEvent(
+			ctx,
+			runJournal,
+			result.RunID,
+			events.EventKindRunCompleted,
+			kinds.RunCompletedPayload{
+				ArtifactsDir:   result.ArtifactsDir,
+				JobsTotal:      len(jobs),
+				JobsSucceeded:  succeeded,
+				JobsFailed:     failed,
+				JobsCancelled:  canceled,
+				DurationMs:     durationMs,
+				ResultPath:     result.ResultPath,
+				SummaryMessage: "completed",
+			},
+		)
+	case runStatusCanceled:
+		reason := strings.TrimSpace(result.Error)
+		if reason == "" {
+			reason = strings.TrimSpace(result.TeardownError)
+		}
+		return submitRunEvent(
+			ctx,
+			runJournal,
+			result.RunID,
+			events.EventKindRunCancelled,
+			kinds.RunCancelledPayload{
+				Reason:     reason,
+				DurationMs: durationMs,
+			},
+		)
+	default:
+		errText := strings.TrimSpace(result.Error)
+		if errText == "" {
+			errText = strings.TrimSpace(result.TeardownError)
+		}
+		return submitRunEvent(
+			ctx,
+			runJournal,
+			result.RunID,
+			events.EventKindRunFailed,
+			kinds.RunFailedPayload{
+				ArtifactsDir: result.ArtifactsDir,
+				DurationMs:   durationMs,
+				Error:        errText,
+				ResultPath:   result.ResultPath,
+			},
+		)
+	}
+}
+
+func closeRunJournal(runJournal *journal.Journal) error {
+	if runJournal == nil {
+		return nil
+	}
+	closeCtx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+	if err := runJournal.Close(closeCtx); err != nil {
+		return fmt.Errorf("close run journal: %w", err)
+	}
+	return nil
+}
+
+func atLeastOne(value int) int {
+	if value < 1 {
+		return 1
+	}
+	return value
+}
