+++ pkg/compozy/runs/run.go
@@ -4,14 +4,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"iter"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
-
-	"github.com/compozy/compozy/pkg/compozy/events"
 )
 
 var (
@@ -151,78 +147,6 @@ func (r *Run) Summary() RunSummary {
 	return r.summary
 }
 
-// Replay yields all events from fromSeq to EOF, tolerating a truncated final line.
-func (r *Run) Replay(fromSeq uint64) iter.Seq2[events.Event, error] {
-	return func(yield func(events.Event, error) bool) {
-		if r == nil {
-			yield(events.Event{}, errors.New("replay run: nil run"))
-			return
-		}
-
-		file, err := os.Open(r.eventsPath)
-		if err != nil {
-			yield(events.Event{}, fmt.Errorf("open run events: %w", err))
-			return
-		}
-		defer func() {
-			_ = file.Close()
-		}()
-
-		partialFinalLine, err := fileHasPartialFinalLine(file)
-		if err != nil {
-			yield(events.Event{}, fmt.Errorf("inspect run events: %w", err))
-			return
-		}
-
-		stopReplay := errors.New("stop replay")
-		stopPendingDecode := errors.New("stop pending decode")
-		var pendingDecodeErr error
-		var pendingLine int
-
-		readErr := forEachEventLine(file, func(rawLine []byte, lineNumber int) error {
-			line := bytesTrimSpace(rawLine)
-			if len(line) == 0 {
-				return nil
-			}
-
-			ev, err := decodeEventLine(line, lineNumber)
-			if err != nil {
-				pendingDecodeErr = err
-				pendingLine = lineNumber
-				return nil
-			}
-			if pendingDecodeErr != nil {
-				return stopPendingDecode
-			}
-			if ev.Seq < fromSeq {
-				return nil
-			}
-			if !yield(ev, nil) {
-				return stopReplay
-			}
-			return nil
-		})
-		switch {
-		case errors.Is(readErr, stopReplay):
-			return
-		case errors.Is(readErr, stopPendingDecode):
-			yield(events.Event{}, pendingDecodeErr)
-			return
-		case readErr != nil:
-			yield(events.Event{}, fmt.Errorf("scan run events: %w", readErr))
-			return
-		}
-		if pendingDecodeErr == nil {
-			return
-		}
-		if partialFinalLine {
-			yield(events.Event{}, fmt.Errorf("%w: line %d", ErrPartialEventLine, pendingLine))
-			return
-		}
-		yield(events.Event{}, pendingDecodeErr)
-	}
-}
-
 func defaultRunPaths(workspaceRoot, runID string) runPaths {
 	runDir := filepath.Join(runsDirForWorkspace(workspaceRoot), runID)
 	return runPaths{
@@ -290,162 +214,4 @@ func normalizeRunSummary(workspaceRoot, runID, runDir string, record runRecord) 
 		summary.EndedAt = &endedAt
 	}
 	return summary
-}
-
-func deriveRunState(paths runPaths) (derivedRunState, error) {
-	state := derivedRunState{}
-
-	status, err := loadResultStatus(paths.resultPath)
-	if err != nil {
-		return state, err
-	}
-	if status != "" {
-		state.status = status
-	}
-
-	eventState := bestEffortRunStateFromEvents(paths.eventsPath)
-	if state.status == "" {
-		state.status = eventState.status
-	}
-	if state.endedAt == nil {
-		state.endedAt = eventState.endedAt
-	}
-	return state, nil
-}
-
-func loadResultStatus(resultPath string) (string, error) {
-	if strings.TrimSpace(resultPath) == "" {
-		return "", nil
-	}
-	payload, err := os.ReadFile(resultPath)
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return "", nil
-		}
-		return "", fmt.Errorf("read run result: %w", err)
-	}
-
-	var record resultRecord
-	if err := json.Unmarshal(payload, &record); err != nil {
-		return "", fmt.Errorf("decode run result: %w", err)
-	}
-	return normalizeStatus(record.Status), nil
-}
-
-func bestEffortRunStateFromEvents(eventsPath string) derivedRunState {
-	file, err := os.Open(eventsPath)
-	if err != nil {
-		return derivedRunState{}
-	}
-	defer func() {
-		_ = file.Close()
-	}()
-
-	state := derivedRunState{}
-	if err := forEachEventLine(file, func(rawLine []byte, lineNumber int) error {
-		line := bytesTrimSpace(rawLine)
-		if len(line) == 0 {
-			return nil
-		}
-		ev, err := decodeEventLine(line, lineNumber)
-		if err != nil {
-			return err
-		}
-		switch ev.Kind {
-		case events.EventKindRunStarted, events.EventKindRunQueued:
-			if state.status == "" {
-				state.status = defaultRunStatus()
-			}
-		case events.EventKindRunCompleted:
-			state = derivedRunState{status: publicRunStatusCompleted, endedAt: timePointer(ev.Timestamp)}
-		case events.EventKindRunFailed:
-			state = derivedRunState{status: publicRunStatusFailed, endedAt: timePointer(ev.Timestamp)}
-		case events.EventKindRunCancelled:
-			state = derivedRunState{status: publicRunStatusCancelled, endedAt: timePointer(ev.Timestamp)}
-		}
-		return nil
-	}); err != nil {
-		return state
-	}
-	return state
-}
-
-func decodeEventLine(line []byte, lineNumber int) (events.Event, error) {
-	var ev events.Event
-	if err := json.Unmarshal(line, &ev); err != nil {
-		return events.Event{}, fmt.Errorf("decode run event line %d: %w", lineNumber, err)
-	}
-	if err := validateSchemaVersion(ev.SchemaVersion); err != nil {
-		return events.Event{}, err
-	}
-	return ev, nil
-}
-
-func validateSchemaVersion(version string) error {
-	major, _, ok := strings.Cut(strings.TrimSpace(version), ".")
-	if !ok || major == "" {
-		return &SchemaVersionError{Version: version}
-	}
-	expectedMajor, _, _ := strings.Cut(events.SchemaVersion, ".")
-	if major != expectedMajor {
-		return &SchemaVersionError{Version: version}
-	}
-	return nil
-}
-
-func normalizeStatus(status string) string {
-	switch strings.ToLower(strings.TrimSpace(status)) {
-	case "":
-		return ""
-	case "succeeded", publicRunStatusCompleted:
-		return publicRunStatusCompleted
-	case "canceled", publicRunStatusCancelled:
-		return publicRunStatusCancelled
-	default:
-		return strings.TrimSpace(status)
-	}
-}
-
-func defaultRunStatus() string {
-	return publicRunStatusRunning
-}
-
-func isTerminalStatus(status string) bool {
-	switch normalizeStatus(status) {
-	case publicRunStatusCompleted, publicRunStatusFailed, publicRunStatusCancelled:
-		return true
-	default:
-		return false
-	}
-}
-
-func timePointer(value time.Time) *time.Time {
-	if value.IsZero() {
-		return nil
-	}
-	copyValue := value
-	return &copyValue
-}
-
-func fileHasPartialFinalLine(file *os.File) (bool, error) {
-	if _, err := file.Seek(0, io.SeekStart); err != nil {
-		return false, err
-	}
-
-	info, err := file.Stat()
-	if err != nil {
-		return false, err
-	}
-	if info.Size() == 0 {
-		return false, nil
-	}
-
-	var tail [1]byte
-	if _, err := file.ReadAt(tail[:], info.Size()-1); err != nil {
-		return false, err
-	}
-	if _, err := file.Seek(0, io.SeekStart); err != nil {
-		return false, err
-	}
-	return tail[0] != '\n', nil
 }
